# openapi.yaml
openapi: 3.0.0
servers:
  - url: http://localhost8080
info:
  description: Contract API For The Skill-Matrix Application. This API contains all the neccessary endpoints for CRUD operations based on project requirements. In a nutshell, the application for which this API was created, contains all the active employees currently working for SoftwareONE. The purpose of this application is to aggregate all employees in one place so that any manager can access and contact any employee based on skills and other criterias.
  version: "1.0.0"
  title: Skill-Matrix API 
  contact:
    email: zourab.lavrentiadis@softwareone.com
  license:
    name: SoftwareONE
    url: 'http://www.softwareone.com'
...
paths:
...
  /job-title:
    get:
      tags:
        - job-title
      summary: fetch all job-titles from the database
      operationId: getJobTitles
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/JobTitle'
        4XX:
          description: problem caused by the user
        5XX:
          description: problem caused by the server   
  
    post:
      tags:
        - job-title
      summary: add a job-title to the database
      operationId: addJobTitle
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/JobTitle'
      responses:
        201:
           description: created
           content:
            application/json:
              schema:
                $ref: '#/components/schemas/JobTitle'
        4XX:
          description: problem caused by the user
        5XX:
          description: problem caused by the server   
  /job-title/{job-titleId}:
    get:
      tags:
        - job-title
      summary: find job-title by id
      operationId: getJobTitleById
      parameters:
        - name: job-titleId
          in: path
          description: id of job-title to return
          required: true
          schema:
            type: integer
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/JobTitle'
        4XX:
          description: problem caused by the user
        5XX:
          description: problem caused by the server   
 
    put:
      tags:
        - job-title
      summary: update a job-title
      operationId: updateJobTitle
      parameters:
      - in: path
        name: job-titleId
        required: true
        schema:
          type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
               $ref: '#/components/schemas/JobTitle'
      responses:
        201:
          description: created
        4XX:
          description: problem caused by the user
        5XX:
          description: problem caused by the server   
          
    delete:
      tags:
        - job-title
      summary: delete a job-title by id
      operationId: deleteJobTitle
      parameters:
      - in: path
        name: job-titleId
        required: true
        schema:
          type: integer
      responses:
        201:
          description: deleted
        4XX:
          description: problem caused by the user
        5XX:
          description: problem caused by the server      


 #department
  /department:
    get:
      tags:
        - department
      summary: fetch all departments from the database
      operationId: getDepartments
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Department'
        4XX:
          description: problem caused by the user
        5XX:
          description: problem caused by the server   
  
    post:
      tags:
        - department
      summary: add a department to the database
      operationId: addDepartment
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Department'
      responses:
        201:
           description: created
           content:
            application/json:
              schema:
                $ref: '#/components/schemas/Department'
        4XX:
          description: problem caused by the user
        5XX:
          description: problem caused by the server   
  /department/{departmentId}:
    get:
      tags:
        - department
      summary: find a department by id
      operationId: getDepartmentById
      parameters:
        - name: departmentId
          in: path
          required: true
          schema:
            type: integer
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Department'
        4XX:
          description: problem caused by the user
        5XX:
          description: problem caused by the server   
  
    put:
      tags:
        - department
      summary: update a department by id
      operationId: updateDepartment
      parameters:
      - in: path
        name: departmentId
        required: true
        schema:
          type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
               $ref: '#/components/schemas/Department'
      responses:
        201:
          description: item created
        4XX:
          description: problem caused by the user
        5XX:
          description: problem caused by the server   
          
    delete:
      tags:
        - department
      summary: delete a department by id
      operationId: deleteDepartment
      parameters:
      - in: path
        name: departmentId
        required: true
        schema:
          type: integer
      responses:
        201:
          description: deleted
        4XX:
          description: problem caused by the user
        5XX:
          description: problem caused by the server      


 #category
...
components:
  schemas:
...
    JobTitle:
      type: object
      properties:
        jobTitleId:
          type: integer
          example: 1
        jobTitle:
          type: string
          enum:
            -  developer
            -  manager
            -  administrator 
    Department:
      type: object
      properties:
        departmentId:
          type: integer
          example: 1
        department:
          type: string
          enum:
            - application
            - sales
            - marketing  
...